# Iteratively traverses the document during IonValue.makeReadOnly