*Description of changes:*

Applies the iterative technique in #542 to `IonValue.makeReadOnly`. 

### Performance

Datasets:
* entry.ion - 225 KB, containing deeply-nested structs with recurring structure, short- and medium-length strings, and numeric values.
* committed.10n - 140 B, single top-level struct with local symbol table.
* log.10n - 22MB, containing nested structs with recurring structure, symbols, annotations, and numeric values.

#### IonValue.makeReadOnly

| dataset                | before (us/op) | after (us/op)       |
| -----------------| ---------------|-----------------|
| entry.ion              |  0.759 	        | 0.775 (+2.10)  |
| committed.10n    |  0.762             | 0.758 (-0.52%)   |
| log.10n                 | 0.848              | 0.858 (+1.17%) |