@desaikd 2023-08-21
> The performance table in the PR overview mentions `IonValue.hashCode`. Is that just a typo?

Yes, thats a Typo. Just changed it.