@tgregg 2023-08-23
I pulled this down locally to start building on top of it and noticed that IonSystemLiteTest was running very slowly (~minutes, where previously it would finish in less than a second). During my investigation I started by unifying the structure of the method with that of containerHashCode to make sure nothing was wrong in the iteration logic. That is the bulk of the diff in this change. When that didn't fix it, I eventually realized that we should short-circuit on container values (i.e., not step into them) unless their isSymbolIdPresent flag is true. This got IonSystemLiteTest running in under a second again. I have one remaining TODO in the code to check the equivalent of IonSystemLiteTest for `makeReadOnly`. I'll look into that tomorrow.